refactor : 문제 리스트 폴더 구조 변경(query -> entities)
