+++ src/entities/problems/model/types.ts
